feat(api): megaparse under api (#93)
* feat(api): megaparse under api

This commit updates the dependencies in the `pyproject.toml` file. It adds new packages such as `fastapi`, `uvicorn`, `ratelimit`, `requests`, `backoff`, `pypdf`, `psutil`, and `numpy`. Additionally, it updates the version of `pycryptodome` to 3.21.0.

* feat(api): Add endpoint for URL upload

This commit adds a new endpoint `/url` to the API that allows users to upload a URL and retrieve the extracted content from the webpage. The endpoint uses the `PlaywrightURLLoader` from the `langchain_community.document_loaders` module to load the webpage and extract the content. The extracted content is then returned as a response.

* feat(api): Add support for PDF file upload via URL

* Re facto Megaparse & API

* Refacto API & fix import conflict

* fix evaluation script & Read me

* fix readme

* add tests

* fix: add fake models for tests

* Refactor upload_file function to allow for dynamic model selection

* fix: PR comments

* fix: PR comments

* fix: PR comments

* fix: PR comments

* fix: PR comments

* fix: reorganize files and fix PR comments

* fix: readme

---------

Co-authored-by: chloedia <chloedaems0@gmail.com>