feat: Add GitHub Actions workflow for building Docker images (#190)
* feat: Add GitHub Actions workflow for building Docker images

This commit adds a new GitHub Actions workflow file, "test-build-docker.yml", which is responsible for building Docker images. The workflow is triggered on pull requests to the main branch. It uses the "ubuntu-latest" runner and supports building both regular and GPU-enabled Docker images by utilizing the "Dockerfile" and "Dockerfile.gpu" files respectively. The workflow also sets up QEMU and Docker Buildx for the build process. The resulting Docker images are not pushed to a registry but are cached for future builds.

Refs: #187

* feat: Add GitHub Actions workflow for testing and building Docker images